perbaiki show on menu

+++ database/migrations/2025_11_22_122015_add_show_on_menu_to_posts_table.php
@@ -14,14 +14,18 @@ return new class extends Migration
     public function up()
     {
         Schema::table('posts', function (Blueprint $table) {
-            $table->string('show_on_menu')->default('Show');
+            if (!Schema::hasColumn('posts', 'show_on_menu')) {
+                $table->string('show_on_menu')->default('Show');
+            }
         });
     }
 
     public function down()
     {
         Schema::table('posts', function (Blueprint $table) {
-            $table->dropColumn('show_on_menu');
+            if (Schema::hasColumn('posts', 'show_on_menu')) {
+                $table->dropColumn('show_on_menu');
+            }
         });
     }
 };
